scripts/antibiotic-susceptibility/Aim 4/AIM4/Embedding_Pilot_Exp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncer==0.0.8
asyncio==4.0.0
attrs==25.3.0
backoff==2.2.1
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==2.6.27
dspy-ai==2.6.27
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json_repair==0.44.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
litellm==1.76.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.16
numpy==2.0.2
openai==1.102.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
propcache==0.3.2
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
